refactor: Extract MonotonicityMixin from hjb_gfdm.py (#389)
Major refactoring of HJB GFDM solver to improve maintainability:

1. Extract MonotonicityMixin (hjb_gfdm_monotonicity.py - 709 lines)
   - _solve_monotone_constrained_qp() - Main QP solving entry point
   - _solve_unconstrained_fallback() - SVD/normal equations fallback
   - _check_monotonicity_violation() - Detect when QP is needed
   - _check_m_matrix_property() - Verify FD weights satisfy M-matrix
   - _build_monotonicity_constraints() - Indirect Taylor-based constraints
   - _build_hamiltonian_gradient_constraints() - Direct dH/du_j >= 0
   - _build_coefficient_bounds() - Physics-based coefficient bounds
   - _build_boundary_stability_constraints() - Boundary point constraints
   - print_qp_diagnostics() - QP statistics reporting
   - _compute_fd_weights_from_taylor() - FD weight extraction

2. Refactor grid<->collocation mapping
   - Replace placeholder implementations with proper interpolation
   - Use existing particle/interpolation.py utilities
   - interpolate_grid_to_particles() for grid->collocation (RegularGridInterpolator)
   - interpolate_particles_to_grid() for collocation->grid (RBF)

3. Extend GFDMOperator with boundary condition support
   - Add boundary_indices, domain_bounds, boundary_type parameters
   - Add _create_ghost_particles() for no-flux BC
   - Add _get_neighbor_values() for ghost particle handling

4. Update FPGFDMSolver to accept BC parameters

Result: hjb_gfdm.py reduced from 2340 -> 1450 lines (-38%)
